Upload working csv
Version 1.007 Now that we are semi functional all major additions will get a main vers and mini ver

so 1.007 will be version 1 iteration 7 The only way to move up a mini iteration is to fix a bug or add functionality. We will also be tracking this so we know what worked on what iterations.